I'll make it work. (#3)
* Fixed crash

* A bit of collision detection & response created.

* Ceiling collision done

* Finally human physics

* Human physics actually done.

* Animation work in progress

* Stil WIP

* Still work in progress, but almost done I think.

* A

* GG u know why

* Fixed a platform bug.

* Character flips when moving

* Should be better this way.

+++ Application/Source/Renderer/Shader.cpp
@@ -39,7 +39,7 @@ auto Shader::ReadFile(const std::string &file_path) -> std::string
     if (in)
     {
         in.seekg(0, std::ios::end);
-        result.resize(in.tellg());
+        result.resize(static_cast<const unsigned int>(in.tellg()));
         in.seekg(0, std::ios::beg);
         in.read(&result[0], result.size());
         in.close();
